requirements.txt
selenium
webdriver-manager
python-dotenv
openpyxl
requests
Pillow
